streamlit_app.py

import streamlit as st
from PIL import Image
import os

# Title
st.markdown("<h1 style='color: #00FF88;'>SnapTrade App</h1>", unsafe_allow_html=True)

# Persistent ticker selection
ticker = st.selectbox("Select Ticker Symbol", ["MNQ", "NQ", "QQQ", "SPY", "AAPL", "BTCUSD"], index=0)
st.session_state["ticker"] = ticker

# File uploader
uploaded_file = st.file_uploader("Upload a TradingView screenshot", type=["png", "jpg", "jpeg"])

# Trade suggestion display
if uploaded_file:
    image = Image.open(uploaded_file)
    st.image(image, caption="Uploaded Chart", use_column_width=True)

    # Placeholder logic (this should be replaced with actual analysis later)
    st.subheader("Trade Plan Suggestion")
    st.markdown("**Confidence Level:** 92%")
    st.markdown("**Direction:** Long")
    st.markdown("**Entry Price:** 21413.00")
    st.markdown("**Stop Loss:** 21378.75")
    st.markdown("**Take Profit:** 21460.00")

    # Feedback
    st.markdown("### Did this trade plan work for you?")
    col1, col2 = st.columns(2)
    with col1:
        st.button("✅ Yes")
    with col2:
        st.button("❌ No")

    # ChatGPT explanation
    st.markdown("#### AI Explanation")
    st.write("This trade plan was based on a bullish signal where price bounced above the 50 SMA with increasing volume and RSI crossing 50. Confidence level exceeded 75%, triggering the suggestion.")

