requirements.txt
streamlit
tensorflow==2.15.0
openai
pillow
scikit-learn
